web3js: Update example/contracts
